MONAI Application Package Specification
Adds an incomplete, first draft of the MONAI Application Package (MAP) specification.

This initial commits contains a description of what a MAP is and is intended to achieve, and an initial set of requirements.

This commit does not include any design or implementation. These will be made available in future commits.