map: Improve specification based on Packager development
This change adds the concept of expected working directory for the application.

Additionally, the input and output paths are now expected to be relative to the working directory.

The primary reason for this is to better enable local (not containerized) application development. When the application uses local file-system paths it doesn't matter which path the application is executed in when not containerized; and enables a no code change operation when containerizing the application.

Additionally, formatting details have been added for version types. We now specify that we expect semantic version values.